+++ guillotina_nuclia/tests/test_content.py
@@ -73,6 +73,66 @@ async def test_api(guillotina):
 
     response, status = await guillotina(
         "POST",
+        "/db/guillotina/@NucliaPredictStateless",
+        data=json.dumps(
+            {
+                "question": "Foo question",
+                "history": [
+                    {"author": "USER", "text": "Foo question"},
+                    {
+                        "author": "NUCLIA",
+                        "text": "Not enough data to answer this.",
+                    },
+                    {"author": "USER", "text": "Foo question 2"},
+                    {
+                        "author": "NUCLIA",
+                        "text": "Not enough data to answer this.",
+                    },
+                ],
+                "context": ["some context"],
+            }
+        ),
+    )
+    assert status == 200
+    assert response["answer"] == "Not enough data to answer this."
+    assert response["history"] == [
+        {"author": "USER", "text": "Foo question"},
+        {"author": "NUCLIA", "text": "Not enough data to answer this."},
+        {"author": "USER", "text": "Foo question 2"},
+        {"author": "NUCLIA", "text": "Not enough data to answer this."},
+        {"author": "USER", "text": "Foo question"},
+        {"author": "NUCLIA", "text": "Not enough data to answer this."},
+    ]
+    assert response["response"]["answer"] == "Not enough data to answer this."
+
+    response, status = await guillotina(
+        "POST",
+        "/db/guillotina/@NucliaPredictStatelessStream",
+        data=json.dumps(
+            {
+                "question": "Foo question",
+                "history": [
+                    {"author": "USER", "text": "Foo question"},
+                    {
+                        "author": "NUCLIA",
+                        "text": "Not enough data to answer this.",
+                    },
+                    {"author": "USER", "text": "Foo question 2"},
+                    {
+                        "author": "NUCLIA",
+                        "text": "Not enough data to answer this.",
+                    },
+                ],
+                "context": ["some context"],
+            }
+        ),
+    )
+    assert status == 200
+    assert response.startswith(b"Not enough data to answer this.")
+    assert b'"type": "end"' in response
+
+    response, status = await guillotina(
+        "POST",
         "/db/guillotina/@NucliaAskStream",
         data=json.dumps({"question": "Foo question"}),
     )